Move to hook-utils
